move newrelic integration from subpackage to main package (#155)
* move newrelic integration from subpackage to main package

* clean up imports
