+++ efopen/newrelic_executor.py
@@ -1,10 +1,9 @@
 from copy import deepcopy
 import logging
 
-from efopen.ef_config import EFConfig
-from efopen.ef_utils import kms_decrypt
-
-from interface import NewRelic, AlertPolicy
+from ef_config import EFConfig
+from ef_utils import kms_decrypt
+from newrelic_interface import NewRelic, AlertPolicy
 
 
 logging.basicConfig(level=logging.INFO)
